% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddict.R
\name{add_sys_words}
\alias{add_sys_words}
\title{Add or remove words from list}
\usage{
add_sys_words(dict, words, freq, tags)
}
\arguments{
\item{dict}{a list}

\item{words}{words to add or remove}

\item{freq}{frequencies for words}

\item{tags}{POS tags}
}
\description{
Add or remove words from list
}
