% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddict.R
\name{load_sys_dict}
\alias{load_sys_dict}
\title{Load system dictionaries}
\usage{
load_sys_dict(filePath)
}
\arguments{
\item{filePath}{jiebaR dict path}
}
\description{
This function can load a jiebaR system dictionary, and then we can use remove_sys_words() and add_sys_words() to edit the dictionary.
}
\examples{
\dontrun{
test = load_sys_dict(jiebaR::DICTPATH)
}
}

