% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddict.R
\name{load_user_dict}
\alias{load_user_dict}
\title{Load user dictionaries}
\usage{
load_user_dict(filePath, default_tag = "n")
}
\arguments{
\item{filePath}{jiebaR dict path}

\item{default_tag}{default frequency}
}
\description{
This function can load a jiebaR user dictionary, and then we can use remove_user_words() and add_user_words() to edit the dictionary.
}
\examples{
\dontrun{
test = load_user_dict(jiebaR::USERPATH)
}
}

