% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loaddict.R
\name{loaddict}
\alias{loaddict}
\title{LoadDIctionaries}
\usage{
loaddict(filePath, defaultWeight, defaultTag)
}
\arguments{
\item{filePath}{SCEL file path}

\item{defaultWeight}{output path}

\item{defaultTag}{default frequency}
}
\description{
This function can load jiebaR dictionaries
}
\examples{
\dontrun{
decode_scel(scel = "test.scel",output = "test.dict",freq = 1)
}
}

