% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loaddict.R
\name{remove_user_words}
\alias{add_user_words}
\alias{remove_user_words}
\title{Add or remove words from list}
\usage{
remove_user_words(dict, words)

add_user_words(dict, words, tags)
}
\arguments{
\item{dict}{a list}

\item{words}{words to add or remove}

\item{tags}{POS tags}
}
\description{
add or remove words from list
}
\examples{
\dontrun{
remove_user_words(dict,words)
}
}

