% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddict.R
\name{remove_words}
\alias{remove_words}
\title{Add or remove words from list}
\usage{
remove_words(dict, words)
}
\arguments{
\item{dict}{a list}

\item{words}{words to add or remove}
}
\description{
add or remove words from list
}
\examples{
\dontrun{
remove_words(dict,words)
}
}

