% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddict.R
\name{write_dict}
\alias{write_dict}
\title{Write the loaded dictionaries to a file}
\usage{
write_dict(dict, output, cpp = TRUE, cpp_progress = TRUE)
}
\arguments{
\item{dict}{a list}

\item{output}{output path}

\item{cpp}{use cpp mode, FALSE will give better error message}

\item{cpp_progress}{use progress in cpp mode}
}
\description{
Write the loaded dictionaries to a file
}
\examples{
\dontrun{
sysd = load_sys_dict(jiebaR::DICTPATH)
write_dict(sysd,"somepath")
}
}
