% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{decode_scel}
\alias{decode_scel}
\title{Decode SCEL files}
\usage{
decode_scel(scel, output = NULL, freq = 1)
}
\arguments{
\item{scel}{SCEL file path}

\item{output}{output path}

\item{freq}{default frequency}
}
\description{
This function can decode SCEL file into jiebaR dictionaries
}
\examples{
\dontrun{
decode_scel(scel = "test.scel",output = "test.dict",freq = 1)
}
}

