% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{decode_scel}
\alias{decode_scel}
\title{Decode SCEL files}
\usage{
decode_scel(scel, output = NULL, tag = "n", cpp = T, progress = F,
  rdebug = FALSE)
}
\arguments{
\item{scel}{SCEL file path}

\item{output}{output path}

\item{tag}{default tag}

\item{cpp}{use Rcpp}

\item{progress}{TRUE}

\item{rdebug}{display debug info}
}
\description{
This function can decode SCEL file into jiebaR dictionaries
}
\examples{
\dontrun{
decode_scel(scel = "test.scel",output = "test.dict",tag = 1)
}
}
